+++ libimagannotation/src/annotateable.rs
